+++ x/pylons/msgs/create_recipe.go
@@ -24,7 +24,7 @@ func NewMsgCreateRecipe(recipeName, cookbookID, description string,
 	coinInputs types.CoinInputList,
 	itemInputs types.ItemInputList,
 	entries types.WeightedParamList,
-	blockInterval int,
+	blockInterval int64,
 	sender sdk.AccAddress) MsgCreateRecipe {
 	return MsgCreateRecipe{
 		RecipeName:    recipeName,
